weird platform-specific overflow issue in int/long conversion.

+++ src/main/java/com/acrescrypto/zksync/fs/zkfs/ArchiveAccessor.java
@@ -214,7 +214,7 @@ public class ArchiveAccessor {
 	
 	public long timeSlice(int index) {
 		// TODO: (refactor) Eliminate all references to System.currentTimeMillis() and replace with something we can more easily control in tests.
-		return (long) TEMPORAL_SEED_KEY_INTERVAL_MS * index;
+		return (long) TEMPORAL_SEED_KEY_INTERVAL_MS * (long) index;
 	}
 
 	public ZKMaster getMaster() {
